holds: add interactive hold diagram
The data file gains a new domain-specific JSON sublanguage for
describing hold annotations, supporting rounded rectangular holds,
start/end tag annotations, and masking. An SVG renderer is included on
the route page for routes with hold data. This language is of course
subject to change, especially re: names.

I created this `holds` data by opening the full route image in GIMP, and
using the rectangle tool to make selections, at which point I can
directly read off the numeric dimensions from the sidebar. This wasn't
too bad, but it definitely wants a little GUI for drawing on an image
directly and generating the JSON from that.

I'm particularly pleased with how the masking turned out.

The hover behavior works on desktop and mobile: tested Firefox and
Chrome on a Pixel 4a (5G) running Android 13.

Integration into the routes page is totally a hack for now: I made no
effort to position things well, and haven't ported `FadingImage` to an
SVG context (though I have ideas). All to be improved.