holds: improve fading accuracy and reduce jank
This fixes the jank wherein mousing out from a hold would cause the
shade to start fading out but the hold to immediately be removed from
the shade mask, where really it should stick around until the transition
finishes.

I tried a different approach that allowed cleanly fading between all
different holds, instead of only fading the shade but jumping between
holds. It works (it's tricky to get right; I used a map over all the
holds inside the mask definition where the `hoveredHold && ...` is now)
but it needed to animate `opacity` rather than `fill-opacity`, and on
mobile browsers (both Firefox and Chrome) this is janky and laggy. This
approach is arguably less pretty, but feels much snappier on mobile.