holds: pull `strokeWidth` up to top level
As usual, extending the manual DSL in the JSON file for now. Think I'll
eventually have this get parsed from Markdown: emphasis and paragraphs
with the obvious semantics, and hold references as links to fragments:

    ...right hand on the [right start hold](#3)...

Then any bog-standard CommonMark parser or sufficiently motivated pile of
regexes should be able to do the job.