router: fix navigation when initial load is non-root
The repro was to navigate to `/`, click a link to a route page (say, for
route #123), then refresh the page and press Back. This will load the
app with an `initialPath` of `/routes/123`, and then dispatch a
`popstate` event with no previous state. But it should go to whatever
the previous page is, not the initial page.

We can fix this by reading the desired pathname from `window.location`,
which has already been updated. This feels more reliable, anyway, and
now we don't need to manually shuttle paths into our history events.

There's a new potential error case if this would cause us to navigate
past the application root. You can trigger this manually by visiting the
home page, executing `history.pushState(null, "", "/wat/")`, clicking
any in-app link, and pressing Back. I'm not aware of a way to trigger
this other than through manual munging, but in any case the fallback
seems to work nicely.