gallery: keep filtered routes in DOM, just invisible
When toggling back on black routes, this reduces jank on Android from
250ms lag to 100ms lag in dev, or 175ms to 50ms in prod. That's still
more than I'd like, and qualitatively it still *feels* bad, but it's at
least more tolerable.

I think part of the improvement is that `FadingImage`'s initial render
is expensive due to the setter for the `HTMLImageElement.src` setter,
which was showing up as taking 60ms of the dev profile.