data: add climbs 25 through 36
These are the red climbs at Vital Brooklyn. This is mostly
straightforward, though route #28 (Vital Red #4) has two photos because
of how curved it is, so I'll see if I can add some support for getting
the other photo up. Uploaded that secondary image as `0028b` for now,
but it's unused.